online_store_inventory/requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
